Refatorando a ProfileScreen
