--- src/screens/ProfileScreen.tsx
@@ -1,186 +1,0 @@
-// ====== IMPORTS DE DEPENDÊNCIAS E TIPOS ======
-import React from "react";
-import styled from "styled-components/native";
-import { Button, ListItem } from "react-native-elements";
-import { useAuth } from "../contexts/AuthContext";
-import { useNavigation } from "@react-navigation/native";
-import { NativeStackNavigationProp } from "@react-navigation/native-stack";
-import { RootStackParamList } from "../types/navigation";
-import theme from "../styles/theme";
-import Header from "../components/Header";
-import { ViewStyle } from "react-native";
-
-// ====== TIPAGEM DE PROPS ======
-type ProfileScreenProps = {
-  navigation: NativeStackNavigationProp<RootStackParamList, "Profile">;
-};
-
-// ====== COMPONENTE PRINCIPAL ======
-const ProfileScreen: React.FC = () => {
-  // ====== HOOKS ======
-  const { user, signOut } = useAuth();
-  const navigation = useNavigation<ProfileScreenProps["navigation"]>();
-
-  // ====== FUNÇÃO AUXILIAR ======
-  const getRoleText = (role: string) => {
-    switch (role) {
-      case "admin":
-        return "Administrador";
-      case "doctor":
-        return "Médico";
-      case "patient":
-        return "Paciente";
-      default:
-        return role;
-    }
-  };
-
-  // ====== INTERFACE VISUAL ======
-  return (
-    <Container>
-      <Header />
-      <ScrollView contentContainerStyle={styles.scrollContent}>
-        <Title>Meu Perfil</Title>
-
-        {/* ====== INFORMAÇÕES DO USUÁRIO ====== */}
-        <ProfileCard>
-          <Avatar
-            source={{ uri: user?.image || "https://via.placeholder.com/150" }}
-          />
-          <Name>{user?.name}</Name>
-          <Email>{user?.email}</Email>
-          <RoleBadge role={user?.role || ""}>
-            <RoleText>{getRoleText(user?.role || "")}</RoleText>
-          </RoleBadge>
-
-          {user?.role === "doctor" && (
-            <SpecialtyText>Especialidade: {user?.specialty}</SpecialtyText>
-          )}
-        </ProfileCard>
-
-        {/* ====== BOTÕES DE AÇÃO ====== */}
-        <Button
-          title="Voltar"
-          onPress={() => navigation.goBack()}
-          containerStyle={styles.button as ViewStyle}
-          buttonStyle={styles.buttonStyle}
-        />
-
-        <Button
-          title="Editar Perfil"
-          onPress={() => navigation.navigate("EditProfile" as any)}
-          containerStyle={styles.button as ViewStyle}
-          buttonStyle={styles.editButton}
-        />
-
-        <Button
-          title="Sair"
-          onPress={signOut}
-          containerStyle={styles.button as ViewStyle}
-          buttonStyle={styles.logoutButton}
-        />
-      </ScrollView>
-    </Container>
-  );
-};
-
-// ====== ESTILOS JAVASCRIPT ======
-const styles = {
-  scrollContent: {
-    padding: 20,
-  },
-  button: {
-    marginBottom: 20,
-    width: "100%",
-  },
-  buttonStyle: {
-    backgroundColor: theme.colors.primary,
-    paddingVertical: 12,
-  },
-  logoutButton: {
-    backgroundColor: theme.colors.error,
-    paddingVertical: 12,
-  },
-  editButton: {
-    backgroundColor: theme.colors.success,
-    paddingVertical: 12,
-  },
-};
-
-// ====== COMPONENTES STYLED ======
-const Container = styled.View`
-  flex: 1;
-  background-color: ${theme.colors.background};
-`;
-
-const ScrollView = styled.ScrollView`
-  flex: 1;
-`;
-
-const Title = styled.Text`
-  font-size: 24px;
-  font-weight: bold;
-  color: ${theme.colors.text};
-  margin-bottom: 20px;
-  text-align: center;
-`;
-
-const ProfileCard = styled.View`
-  background-color: ${theme.colors.background};
-  border-radius: 8px;
-  padding: 20px;
-  margin-bottom: 20px;
-  align-items: center;
-  border-width: 1px;
-  border-color: ${theme.colors.border};
-`;
-
-const Avatar = styled.Image`
-  width: 120px;
-  height: 120px;
-  border-radius: 60px;
-  margin-bottom: 16px;
-`;
-
-const Name = styled.Text`
-  font-size: 20px;
-  font-weight: bold;
-  color: ${theme.colors.text};
-  margin-bottom: 8px;
-`;
-
-const Email = styled.Text`
-  font-size: 16px;
-  color: ${theme.colors.text};
-  margin-bottom: 8px;
-`;
-
-const RoleBadge = styled.View<{ role: string }>`
-  background-color: ${(props: { role: string }) => {
-    switch (props.role) {
-      case "admin":
-        return theme.colors.primary + "20";
-      case "doctor":
-        return theme.colors.success + "20";
-      default:
-        return theme.colors.secondary + "20";
-    }
-  }};
-  padding: 4px 12px;
-  border-radius: 4px;
-  margin-bottom: 8px;
-`;
-
-const RoleText = styled.Text`
-  color: ${theme.colors.text};
-  font-size: 14px;
-  font-weight: 500;
-`;
-
-const SpecialtyText = styled.Text`
-  font-size: 16px;
-  color: ${theme.colors.text};
-  margin-top: 8px;
-`;
-
-export default ProfileScreen;
