+++ src/screens/ProfileScreen/hooks/useProfileAuth.ts
@@ -1,0 +1,2 @@
+import { useAuth } from "../../../contexts/AuthContext";
+export const useProfileAuth = () => useAuth();
