+++ src/screens/ProfileScreen/models/User.ts
@@ -1,0 +1,7 @@
+export type User = {
+  name: string;
+  email: string;
+  image?: string;
+  role: string;
+  specialty?: string;
+};
