+++ src/screens/ProfileScreen/services/roleService.ts
@@ -1,0 +1,12 @@
+export const getRoleText = (role: string) => {
+  switch (role) {
+    case "admin":
+      return "Administrador";
+    case "doctor":
+      return "Médico";
+    case "patient":
+      return "Paciente";
+    default:
+      return role;
+  }
+};
